Avoid OOB access in q6core

+++ sound/soc/msm/qdsp6v2/q6core.c
@@ -87,7 +87,7 @@ static struct generic_get_data_ *generic_get_data;
 static int parse_fwk_version_info(uint32_t *payload, uint16_t payload_size)
 {
 	size_t ver_size;
-	int num_services;
+	uint16_t num_services;
 
 	pr_debug("%s: Payload info num services %d\n",
 		 __func__, payload[4]);
